# openapi.yaml
openapi: 3.1.0
info:
  title: Telemetry API
  version: 0.8.10
  description: Pre-SIEM enrichment for NetFlow/IPFIX and Zeek with GeoIP, ASN, Threat Intel, and Risk scoring.
servers:
  - url: http://localhost/v1
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    IngestResultMultiStatus:
      type: object
      properties:
        status: { type: string, enum: [partial, ok, error] }
        succeeded: { type: integer }
        failed: { type: integer }
        errors:
          type: array
          items:
            type: object
            properties:
              index: { type: integer }
              error: { type: string }
...
paths:
...
  /ingest/bulk:
    post:
      summary: Bulk ingest with type hint
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [type, records]
              properties:
                type: { type: string, enum: [zeek, netflow] }
                records:
                  type: array
                  items: { type: object }
      responses:
        '200': { description: OK }
        '207':
          description: Partial errors
          content:
            application/json:
              schema: { $ref: '#/components/schemas/IngestResultMultiStatus' }
...
  /indicators:
    put:
      summary: Upsert threat indicators
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [ip_or_cidr, category, confidence]
              properties:
                ip_or_cidr: { type: string }
                category: { type: string }
                confidence: { type: integer, minimum: 0, maximum: 100 }
      responses:
        '200': { description: OK }
  /indicators/{id}:
    delete:
      summary: Delete indicator by ID
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        '200': { description: OK }
...